requirements.txt
# Core Dependencies
torch>=2.0.0
transformers>=4.35.0
openai-whisper>=20231117
langchain>=0.1.0
langchain-community>=0.0.10

# Vector Database & Embeddings
faiss-cpu>=1.7.4
chromadb>=0.4.0
sentence-transformers>=2.2.0

# NLP & Korean Models
kobert-transformers>=0.5.1

# API & Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
pyaudio>=0.2.13
noisereduce>=3.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
aiohttp>=3.9.0
tqdm>=4.66.0

# Evaluation
jiwer>=3.0.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.1.0
